check if vimrc exists

+++ install_vimrc.sh
@@ -3,9 +3,11 @@ DIR=$(cd $(dirname $0); pwd)/vim
 echo $DIR
 
 if vim --version > /dev/null 2>&1; then
-  echo backup: mv ~/.vimrc ~/.vimrc.bk
-  mv ~/.vimrc ~/.vimrc.bk
-  echo install: ln -s ${DIR}/.vimrc ~/.vimrc
+  if [[ -e ~/.vimrc ]]; then
+    echo backup: mv ~/.vimrc ~/.vimrc.bk
+    mv ~/.vimrc ~/.vimrc.bk
+  fi
+  echo install: ln -sf ${DIR}/.vimrc ~/.vimrc
   ln -sf ${DIR}/.vimrc ~/.vimrc
 else
   echo vim not exists...
